Add check for admin role for game endpoints.

+++ api/src/main/java/com/gmw/api/rest/activity/game/DeleteGameActivity.java
@@ -1,27 +1,31 @@
 package com.gmw.api.rest.activity.game;
 
 import com.gmw.api.rest.activity.Activity;
+import com.gmw.api.rest.utils.RoleChecker;
 import com.gmw.services.ServiceManager;
 import com.gmw.services.ServiceManagerFactoryImpl;
 import com.gmw.services.exceptions.ResourceNotDeletedException;
-import com.gmw.services.exceptions.ServiceManagerFactoryException;
 import com.gmw.services.game.DBGameService;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 
+@AllArgsConstructor
 public class DeleteGameActivity extends Activity<Void> {
 
     private final Long id;
-
-    public DeleteGameActivity(Long id) {
-        this.id = id;
-    }
+    private final Long userId;
 
     @Override
     protected Void realExecute() throws ResourceNotDeletedException {
         try (ServiceManager serviceManager = new ServiceManagerFactoryImpl().createSqlServiceManager()) {
-            DBGameService service = serviceManager.getDbGameService();
-            service.deleteGame(id);
-            status = HttpStatus.OK;
+            if (RoleChecker.isAdmin(serviceManager, userId))
+            {
+                DBGameService service = serviceManager.getDbGameService();
+                service.deleteGame(id);
+                status = HttpStatus.OK;
+            } else {
+                setForbidden();
+            }
         } catch (Exception e) {
             throw new ResourceNotDeletedException();
         }
